+++ src/ska_low_mccs_spshw/subrack/subrack_simulator.py
@@ -114,7 +114,7 @@ class SubrackSimulator(SubrackProtocol):
         },
         "power_supply_currents": {
             "length": 2,
-            "default": [2.2, 2.8],
+            "default": [2.2, 2.4],
             "writable": False,
         },
         "power_supply_voltages": {
@@ -154,7 +154,7 @@ class SubrackSimulator(SubrackProtocol):
         # },
         "tpm_voltages": {
             "length": 8,
-            "default": [5.0] * 8,
+            "default": [12.0] * 8,
             "writable": False,
         },
         "board_info": {
@@ -647,7 +647,7 @@ class SubrackSimulator(SubrackProtocol):
                 "V_1V1": 1.1,
                 "V_CORE": 1.2,
                 "V_1V5": 1.5,
-                "V_3V3": 2.3,
+                "V_3V3": 3.3,
                 "V_5V": 5.0,
                 "V_3V": 3.0,
                 "V_2V8": 2.8,