+++ src/ska_low_mccs_spshw/tile/tile_device.py
@@ -1501,8 +1501,10 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         archive_abs_change=0.1,
         min_value=15.0,
         max_value=70.0,
-        min_alarm=16.0,
-        max_alarm=65.0,
+        min_alarm=15.0,
+        max_alarm=70.0,
+        min_warning=16.0,
+        max_warning=65.0,
     )
     def boardTemperature(
         self: MccsTile,
@@ -2745,6 +2747,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC0(self: MccsTile) -> float | None:
         """
@@ -2760,6 +2764,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC1(self: MccsTile) -> float | None:
         """
@@ -2775,6 +2781,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC2(self: MccsTile) -> float | None:
         """
@@ -2790,6 +2798,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC3(self: MccsTile) -> float | None:
         """
@@ -2805,6 +2815,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC4(self: MccsTile) -> float | None:
         """
@@ -2820,6 +2832,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC5(self: MccsTile) -> float | None:
         """
@@ -2835,6 +2849,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC6(self: MccsTile) -> float | None:
         """
@@ -2850,6 +2866,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC7(self: MccsTile) -> float | None:
         """
@@ -2865,6 +2883,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC8(self: MccsTile) -> float | None:
         """
@@ -2880,6 +2900,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC9(self: MccsTile) -> float | None:
         """
@@ -2895,6 +2917,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC10(self: MccsTile) -> float | None:
         """
@@ -2910,6 +2934,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC11(self: MccsTile) -> float | None:
         """
@@ -2925,6 +2951,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC12(self: MccsTile) -> float | None:
         """
@@ -2940,6 +2968,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC13(self: MccsTile) -> float | None:
         """
@@ -2955,6 +2985,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC14(self: MccsTile) -> float | None:
         """
@@ -2970,6 +3002,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Celsius",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=90.0,
+        min_alarm=10.0,
     )
     def temperatureADC15(self: MccsTile) -> float | None:
         """
@@ -3685,10 +3719,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         dtype="DevDouble",
         abs_change=0.1,
         archive_abs_change=0.1,
-        min_value=15.0,
-        max_value=75.0,
-        min_alarm=16.0,
-        max_alarm=68.0,
+        min_alarm=10.0,
+        max_alarm=95.0,
     )
     def fpga1Temperature(
         self: MccsTile,
@@ -3704,10 +3736,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         dtype="DevDouble",
         abs_change=0.2,
         archive_abs_change=0.2,
-        min_value=15.0,
-        max_value=75.0,
-        min_alarm=16.0,
-        max_alarm=68.0,
+        min_alarm=10.0,
+        max_alarm=95.0,
     )
     def fpga2Temperature(
         self: MccsTile,
@@ -3937,7 +3967,13 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
             self._attribute_state["ppsDelay"].update(power, post=False)
         return self._attribute_state["ppsDelay"].read()
 
-    @attribute(dtype="DevLong", archive_abs_change=1, abs_change=1)
+    @attribute(
+        dtype="DevLong",
+        archive_abs_change=1,
+        abs_change=1,
+        max_alarm=10,
+        max_warning=4,
+    )
     def ppsDrift(self: MccsTile) -> int:
         """
         Return the observed drift in the ppsDelay of this Tile.
@@ -4569,6 +4605,10 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Amp",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=2.62,
+        max_warning=2.60,
+        min_warning=2.40,
+        min_alarm=2.37,
     )
     def currentFE0(self: MccsTile) -> float | None:
         """
@@ -4584,6 +4624,10 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Amp",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=2.62,
+        max_warning=2.60,
+        min_warning=2.40,
+        min_alarm=2.37,
     )
     def currentFE1(self: MccsTile) -> float | None:
         """
@@ -4599,6 +4643,10 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=2.62,
+        max_warning=2.57,
+        min_warning=2.40,
+        min_alarm=2.37,
     )
     def voltageAVDD3(self: MccsTile) -> float | None:
         """
@@ -4614,6 +4662,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=0.63,
+        min_alarm=0.57,
     )
     def voltageVrefDDR0(self: MccsTile) -> float | None:
         """
@@ -4629,6 +4679,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=0.63,
+        min_alarm=0.57,
     )
     def voltageVrefDDR1(self: MccsTile) -> float | None:
         """
@@ -4654,6 +4706,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=1.26,
+        min_alarm=1.14,
     )
     def voltageMan1V2(self: MccsTile) -> float | None:
         """
@@ -4669,6 +4723,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=0.97,
+        min_alarm=0.83,
     )
     def voltageMGT_AVCC(self: MccsTile) -> float | None:
         """
@@ -4684,6 +4740,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=1.26,
+        min_alarm=1.14,
     )
     def voltageMGT_AVTT(self: MccsTile) -> float | None:
         """
@@ -4699,6 +4757,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=5.19,
+        min_alarm=4.69,
     )
     def voltageMon5V0(self: MccsTile) -> float | None:
         """
@@ -4714,6 +4774,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=3.88,
+        min_alarm=3.51,
     )
     def voltageMon3V3(self: MccsTile) -> float | None:
         """
@@ -4731,6 +4793,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=2.10,
+        min_alarm=1.90,
     )
     def voltageMon1V8(self: MccsTile) -> float | None:
         """
@@ -4748,6 +4812,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=1.16,
+        min_alarm=1.04,
     )
     def voltageSW_AVDD1(self: MccsTile) -> float | None:
         """
@@ -4763,6 +4829,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=2.42,
+        min_alarm=2.18,
     )
     def voltageSW_AVDD2(self: MccsTile) -> float | None:
         """
@@ -4778,6 +4846,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=12.6,
+        min_alarm=11.4,
     )
     def voltageVIN(self: MccsTile) -> float | None:
         """
@@ -4793,6 +4863,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=0.99,
+        min_alarm=0.84,
     )
     def voltageVM_AGP0(self: MccsTile) -> float | None:
         """
@@ -4808,6 +4880,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=0.99,
+        min_alarm=0.84,
     )
     def voltageVM_AGP1(self: MccsTile) -> float | None:
         """
@@ -4823,6 +4897,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=0.99,
+        min_alarm=0.84,
     )
     def voltageVM_AGP2(self: MccsTile) -> float | None:
         """
@@ -4838,6 +4914,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=0.99,
+        min_alarm=0.84,
     )
     def voltageVM_AGP3(self: MccsTile) -> float | None:
         """
@@ -4853,6 +4931,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=0.99,
+        min_alarm=0.84,
     )
     def voltageVM_AGP4(self: MccsTile) -> float | None:
         """
@@ -4868,6 +4948,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=0.99,
+        min_alarm=0.84,
     )
     def voltageVM_AGP5(self: MccsTile) -> float | None:
         """
@@ -4883,6 +4965,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=0.99,
+        min_alarm=0.84,
     )
     def voltageVM_AGP6(self: MccsTile) -> float | None:
         """
@@ -4898,6 +4982,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=0.99,
+        min_alarm=0.84,
     )
     def voltageVM_AGP7(self: MccsTile) -> float | None:
         """
@@ -4913,6 +4999,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=3.56,
+        min_alarm=3.04,
     )
     def voltageVM_CLK0B(self: MccsTile) -> float | None:
         """
@@ -4928,6 +5016,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=3.56,
+        min_alarm=3.04,
     )
     def voltageVM_CLK1B(self: MccsTile) -> float | None:
         """
@@ -4943,6 +5033,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=0.65,
+        min_alarm=0.55,
     )
     def voltageVM_DDR0_VTT(self: MccsTile) -> float | None:
         """
@@ -4958,6 +5050,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=1.30,
+        min_alarm=1.10,
     )
     def voltageVM_DDR1_VDD(self: MccsTile) -> float | None:
         """
@@ -4973,6 +5067,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=0.65,
+        min_alarm=0.55,
     )
     def voltageVM_DDR1_VTT(self: MccsTile) -> float | None:
         """
@@ -4988,6 +5084,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=1.89,
+        min_alarm=1.71,
     )
     def voltageVM_DRVDD(self: MccsTile) -> float | None:
         """
@@ -5003,6 +5101,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=1.19,
+        min_alarm=1.01,
     )
     def voltageVM_DVDD(self: MccsTile) -> float | None:
         """
@@ -5018,6 +5118,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=3.78,
+        min_alarm=3.22,
     )
     def voltageVM_FE0(self: MccsTile) -> float | None:
         """
@@ -5035,6 +5137,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=3.78,
+        min_alarm=3.22,
     )
     def voltageVM_FE1(self: MccsTile) -> float | None:
         """
@@ -5052,6 +5156,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=1.94,
+        min_alarm=1.66,
     )
     def voltageVM_MGT0_AUX(self: MccsTile) -> float | None:
         """
@@ -5067,6 +5173,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=1.94,
+        min_alarm=1.66,
     )
     def voltageVM_MGT1_AUX(self: MccsTile) -> float | None:
         """
@@ -5082,6 +5190,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=3.56,
+        min_alarm=3.04,
     )
     def voltageVM_PLL(self: MccsTile) -> float | None:
         """
@@ -5097,6 +5207,8 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         unit="Volt",
         abs_change=0.1,
         archive_abs_change=0.1,
+        max_alarm=3.78,
+        min_alarm=3.22,
     )
     def voltageVM_SW_AMP(self: MccsTile) -> float | None:
         """